+++ rust/Aoc2021/src/day06.rs
@@ -1,8 +1,8 @@
 use crate::utils;
 
-type Fish = [i64; 9];
+type Fish = [u64; 9];
 
-pub fn breeding_program(lines: &Vec<String>, days: u32) -> Fish {
+pub fn breeding_program(lines: &Vec<String>, days: u64) -> Fish {
     let mut fish: Fish = [0; 9];
 
     for fish_cycle in lines[0].split(',') {
@@ -30,7 +30,7 @@ pub fn breeding_program(lines: &Vec<String>, days: u32) -> Fish {
     fish
 }
 
-pub fn day6(input: &str) -> (i64, i64) {
+pub fn day6(input: &str) -> (u64, u64) {
     let lines = utils::to_lines(input);
     let part1 = breeding_program(&lines, 80).iter().sum();
     let part2 = breeding_program(&lines, 256).iter().sum();
